moved varable components to util folder
